rework api
A new Component class is added, which provides the base
WidgetPlaceholder wrapper for non-connected components.  That in turn
provides the base for the ConnectedComponent class, which replaces the
previous Component class.

Props now have more similarity to React, and own_props is now available
in map_{dispatch,state}_to_props().